Fixed infor operand, moved DTO
