+++ src/main/java/ch/cern/eam/wshub/core/services/INFOR_OPERATION.java
@@ -157,6 +157,9 @@ public enum INFOR_OPERATION {
     USERDEFINEDTABLE_R("USERDEFINEDTABLE_R"),
     USERDEFINEDTABLE_U("USERDEFINEDTABLE_U"),
     USERDEFINEDTABLE_D("USERDEFINEDTABLE_D"),
+    // MENU HIERARCHY
+    MENU_HIERARCHY_ADD("MENU_HIERARCHY_ADD"),
+    MENU_HIERARCHY_DELETE("MENU_HIERARCHY_DELETE"),
 
     OTHER("OTHER")
     ;
