+++ src/main/java/ch/cern/eam/wshub/core/services/administration/UserGroupMenuService.java
@@ -4,6 +4,7 @@ import ch.cern.eam.wshub.core.annotations.Operation;
 import ch.cern.eam.wshub.core.client.InforContext;
 import ch.cern.eam.wshub.core.interceptors.LogDataReferenceType;
 import ch.cern.eam.wshub.core.services.INFOR_OPERATION;
+import ch.cern.eam.wshub.core.services.administration.entities.MenuSpecification;
 import ch.cern.eam.wshub.core.tools.InforException;
 import ch.cern.eam.wshub.core.tools.Tools;
 
@@ -13,10 +14,10 @@ import java.util.List;
 public interface UserGroupMenuService {
     String MENU_FUNCTION_CODE = "BSFOLD";
 
-    @Operation(logOperation = INFOR_OPERATION.WORKORDER_CREATE, logDataReference1 = LogDataReferenceType.RESULT)
+    @Operation(logOperation = INFOR_OPERATION.MENU_HIERARCHY_ADD, logDataReference1 = LogDataReferenceType.RESULT)
     String addToMenuHierarchy(InforContext context, MenuSpecification node) throws InforException;
 
-    @Operation(logOperation = INFOR_OPERATION.WORKORDER_CREATE, logDataReference1 = LogDataReferenceType.RESULT)
+    @Operation(logOperation = INFOR_OPERATION.MENU_HIERARCHY_DELETE, logDataReference1 = LogDataReferenceType.RESULT)
     String deleteFromMenuHierarchy(InforContext context, MenuSpecification node) throws InforException;
 
     static void validateInputNode(MenuSpecification ms) throws InforException {
@@ -34,44 +35,4 @@ public interface UserGroupMenuService {
             throw Tools.generateFault("User group cannot be empty");
         }
     }
-
-    class MenuSpecification {
-        private List<String> menuPath;
-        private String functionCode;
-        private String forUserGroup;
-
-        public MenuSpecification() {
-        }
-
-        public MenuSpecification(List<String> menuPath, String functionCode, String forUserGroup) {
-            this.menuPath = menuPath;
-            this.functionCode = functionCode;
-            this.forUserGroup = forUserGroup;
-        }
-
-        public List<String> getMenuPath() {
-            return menuPath;
-        }
-
-        public void setMenuPath(List<String> menuPath) {
-            this.menuPath = menuPath;
-        }
-
-        public String getFunctionCode() {
-            return functionCode;
-        }
-
-        public void setFunctionCode(String functionCode) {
-            this.functionCode = functionCode;
-        }
-
-        public String getForUserGroup() {
-            return forUserGroup;
-        }
-
-        public void setForUserGroup(String forUserGroup) {
-            this.forUserGroup = forUserGroup;
-        }
-    }
-
 }
